requirements.txt
diffusers==0.30.2
Pillow==10.4.0
PyYAML==6.0.2
transformers==4.44.2
